znotecommit for description of commit
