+++ builtin/echo.c
@@ -3,17 +3,17 @@
 /*                                                        :::      ::::::::   */
 /*   echo.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mohafnh <mohafnh@student.42.fr>            +#+  +:+       +#+        */
+/*   By: smagniny <santi.mag777@student.42madrid    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/19 14:07:23 by mohafnh           #+#    #+#             */
-/*   Updated: 2023/10/24 03:24:29 by mohafnh          ###   ########.fr       */
+/*   Updated: 2023/10/25 12:33:30 by smagniny         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../include/minshell.h"
+#include "../include/header.h"
 
 // Función para imprimir los argumentos y verificar el uso de -n en el comando
-void starts_with_echo(char **input_str) {
+void    starts_with_echo(char **input_str) {
     int i = 0;
     int flagecho = 1;  // Inicialmente, consideramos que no hay -n
 
@@ -28,8 +28,7 @@ void starts_with_echo(char **input_str) {
     if (flagecho) {
         printf("\n");
     }
-
-	return (input_str);
+	// return (input_str);
 }
 
 // Función para verificar si se usa -n en los argumentos
