+++ builtin/env.c
@@ -3,23 +3,23 @@
 /*                                                        :::      ::::::::   */
 /*   env.c                                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mohafnh <mohafnh@student.42.fr>            +#+  +:+       +#+        */
+/*   By: smagniny <santi.mag777@student.42madrid    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/24 09:36:16 by mohafnh           #+#    #+#             */
-/*   Updated: 2023/10/24 19:08:36 by mohafnh          ###   ########.fr       */
+/*   Updated: 2023/10/25 12:33:50 by smagniny         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../include/minshell.h"
+#include "../include/header.h"
 
-int		env(t_env *env)
-{
-	while (env && env->next != NULL)
-	{
-		ft_putendl_fd(env->value, 1);
-		env = env->next;
-	}
-	if (env)
-		ft_putendl_fd(env->value, 1);
-	return (1);
-}
+// int		env(t_env *env)
+// {
+// 	while (env && env->next != NULL)
+// 	{
+// 		ft_putendl_fd(env->value, 1);
+// 		env = env->next;
+// 	}
+// 	if (env)
+// 		ft_putendl_fd(env->value, 1);
+// 	return (1);
+// }
