+++ builtin/pwd.c
@@ -3,14 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   pwd.c                                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mohafnh <mohafnh@student.42.fr>            +#+  +:+       +#+        */
+/*   By: smagniny <santi.mag777@student.42madrid    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/24 03:41:02 by mohafnh           #+#    #+#             */
-/*   Updated: 2023/10/24 10:07:52 by mohafnh          ###   ########.fr       */
+/*   Updated: 2023/10/25 12:31:40 by smagniny         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../include/minshell.h"
+#include "../include/header.h"
 
 int pwd(char *path)
 {
