+++ maintest.c
@@ -6,7 +6,7 @@
 /*   By: smagniny <santi.mag777@student.42madrid    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/24 19:43:49 by smagniny          #+#    #+#             */
-/*   Updated: 2023/10/24 20:19:18 by smagniny         ###   ########.fr       */
+/*   Updated: 2023/10/25 13:39:35 by smagniny         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,8 @@ static  char    *get_inputline(t_var *var)
 {
 	char			*str;
 
-	ft_printf("%d> ");
+	(void)var;
+	ft_printf("$> ");
 	str = readline(NULL);
 	if (ft_strncmp(str, "clear", 5) == 0)
 	{
@@ -24,12 +25,7 @@ static  char    *get_inputline(t_var *var)
 		clear_history();
 	}
 	if (ft_strncmp(str, "exit", 4) == 0)
-	{
-		doublefree(var->u_tokens);
-		free(var->inputline);
-		free(str);
 		exit(0);
-	}
     add_history(str);
 	return (str);
 }
@@ -56,9 +52,3 @@ int	main(int argc, char **argv, char **envp)
 	}
 	return (0);
 }
-
-
-git branch -m main2 master
-git fetch origin
-git branch -u origin/master master
-git remote set-head origin -a
