+++ parser/lexer.c
@@ -6,23 +6,11 @@
 /*   By: smagniny <santi.mag777@student.42madrid    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/20 15:57:42 by smagniny          #+#    #+#             */
-/*   Updated: 2023/10/24 19:53:21 by smagniny         ###   ########.fr       */
+/*   Updated: 2023/10/25 12:30:12 by smagniny         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../lib/header.h"
-
-int	isingle_operator(char *line, int i)
-{
-	return (line[i] == '|' || (line[i] == '>' && line[(i + 1)] != '>')
-		|| (line[i] == '<' && line[(i + 1)] != '<'));
-}
-
-int	isdouble_operator(char *line, int i)
-{
-	return ((line[i] == '>' && line[(i + 1)] == '>')
-		|| (line[i] == '<' && line[(i + 1)] == '<'));
-}
+#include "../include/header.h"
 
 static	char	*get_next_token(char *inputline, int *start)
 {
@@ -30,7 +18,6 @@ static	char	*get_next_token(char *inputline, int *start)
 	int		i;
 
 	i = 0;
-	// printf("starting point: %d\n", *start);
 	if (inputline == NULL || *inputline == '\0')
 		return (NULL);
 	while (inputline[*start] == ' ' || inputline[*start] == '\t' 
@@ -41,7 +28,6 @@ static	char	*get_next_token(char *inputline, int *start)
 	while (inputline[i] != ' ' && inputline[i] != '\t'
 		&& inputline[i] != '\n' && inputline[i] != '\0')
 	{
-		// printf("iteration loop: %d   |   char: %c\n", i, inputline[i]);
 		if (isdouble_operator(inputline, i))
 		{
 			i += 2;
@@ -51,29 +37,14 @@ static	char	*get_next_token(char *inputline, int *start)
 			break ;
 		i++;
 	}
-	if (*start == i)
-	{
-		if (inputline[i] == '\n' || inputline[i] == '\0')
-			return (NULL);
-		word_token = ft_substr_startend(inputline, *start, i + 1);
-		*start += 1;
-		return (word_token);
-	}
-	else
-	{
-		word_token = ft_substr_startend(inputline, *start, i);
-		// ft_printf("string token: '%s' | start: '%d' | end: '%d'\n", word_token, *start, i);
-		*start = i;
-		return (word_token);
-	}
-	*start = i;
-	return (NULL);
+	word_token = ft_copytoken(inputline, start, &i);
+	return (word_token);
 }
 
 
 static	int	gnt_startpoint(char *inputline, int start)
 {
-	int             i;
+	int	i;
 
 	i = 0;
 	if (inputline == NULL || *inputline == '\0')
@@ -113,12 +84,13 @@ char	**unidentified_tokens(t_var *var)
 	while (start < ft_strlen(var->inputline))
 	{
 		start = gnt_startpoint(var->inputline, start);
-		// printf("Word_token_starting point: %d\n", start);
 		nb_of_tokens++;
 	}
 	var->len_tokens = nb_of_tokens;
 	start = 0;
-	unidentified_tokens = (char **)ft_calloc(nb_of_tokens, sizeof(char *));//proteger
+	unidentified_tokens = (char **)ft_calloc(nb_of_tokens, sizeof(char *));
+	if (!unidentified_tokens)
+		return (NULL);
 	while (++i < nb_of_tokens)
 		unidentified_tokens[i] = get_next_token(var->inputline, &start);
 	return (unidentified_tokens);
